Moved micrometer observation to a separate module
