+++ apm-sniffer/apm-toolkit-activation/apm-toolkit-micrometer-activation/pom.xml
@@ -38,7 +38,7 @@
     <dependencies>
         <dependency>
             <groupId>org.apache.skywalking</groupId>
-            <artifactId>apm-toolkit-micrometer-registry</artifactId>
+            <artifactId>apm-toolkit-micrometer-1.10</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
