+++ apm-sniffer/apm-toolkit-activation/apm-toolkit-micrometer-activation/src/test/resources/META-INF/services/org.apache.skywalking.apm.agent.core.boot.BootService
@@ -16,5 +16,5 @@
 #
 #
 
-org.apache.skywalking.apm.plugin.micrometer.PluginBootService
-org.apache.skywalking.apm.plugin.micrometer.ContextManagerExtendOverrideService
+org.apache.skywalking.apm.toolkit.activation.micrometer.PluginBootService
+org.apache.skywalking.apm.toolkit.activation.micrometer.ContextManagerExtendOverrideService
