+++ docs/menu.yml
@@ -44,8 +44,10 @@ catalog:
             path: "/en/setup/service-agent/java-agent/application-toolkit-meter"
           - name: "Across Thread Solution"
             path: "/en/setup/service-agent/java-agent/application-toolkit-trace-cross-thread"
-          - name: "MicroMeter Registry"
+          - name: "Micrometer Registry"
             path: "/en/setup/service-agent/java-agent/application-toolkit-micrometer"
+          - name: "Micrometer 1.10 Observation"
+            path: "/en/setup/service-agent/java-agent/application-toolkit-micrometer-1.10"
           - name: "Webflux Tracing Assistant APIs"
             path: "/en/setup/service-agent/java-agent/application-toolkit-webflux"
           - name: "Kafka Tracing Assistant APIs"
